refactor: remove React.strictMode
